wip: fix crashes found using driver verifier with all flags selected
ksocket: free all socket resoursces in case we can't allocate
         an IRP for the close message.
         fix bug on disconnect/close we couldn't allocate the memory
         so we shouldn't try to free it.
rbd: write outside of allocated buffer size.
scsi_driver_extension: use ExDeleteResourceLite on failures
driver_extension: free buffer on eresource failure.
util: add guard in case the global information was not allocated.
      do not free request element.
      use disconnect instead of trying to access the socket directly
userspace: deallocate scsi->device if allocated
           use `ExDeleteResourceLite` on failures
           insert the device in the device list only if all operations
           were successful (simplifies removal in case of failure)
           guard checking if the reply/request list is empty
           reply list, do not touch srb status of the element, do not try
           to free the element
scsi_function: do not free element if it should not be completed.

The most critical part of the patch is removing SRB and freeing the queue
element on the request list.
We should not free the element because we use the same element in both lists
and don't have a copy of it. Freeing it in both queues will trigger a double
free or worse.
Also don't attempt to change the status or datalength of the SRB this can
also lead to undefined behaviour because the srb can already be completed.

Signed-off-by: Alin Gabriel Serdean <aserdean@cloudbasesolutions.com>