wip: To be squashed
NBD: add null terminator when logging.
Semaphore: add function WnbdReleaseSemaphore which is guarded by
an try/except block.
DrainDeviceQueue: add parameter to check for SRB completion notification.
This is useful because we only are completing our Reply SRB requests, due
to the fact we are adding the same element to both request and reply queues.
Add eresource for devices to guard socket. This replaces our global connection
guarding.

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@cloudbasesolutions.com>